requirements.txt
fastapi
uvicorn
httpx
python-dotenv
json-log-formatter
pytest
cachetools
gc